requirements.txt
Django>=3.2.0
pandas
django-background-tasks
mysqlclient
